// resources/js/types/cuti.d.ts
import { Media } from '.';
import { User } from './user';

export type Cuti = {
  id: number;
  media: Media[];
  user: User;
  media: Media[];
  tgl_pengajuan: string;
  media: Media[];
  tgl_mulai: string;
  media: Media[];
  tgl_selesai: string;
  media: Media[];
  jumlah_hari: string;
  media: Media[];
  alasan: string;
  media: Media[];
  jenis_cuti: string;
  media: Media[];
  approved_by: User;
  media: Media[];
  approved_at: string;
  media: Media[];
  approval_status: string;
  created_at: string;
  updated_at: string;
};